Allow babel transformation to parallelized
When the `ember-cli-babel.throwUnlessParallelizable` flag was enabled, this addon would cause an error to be raised like

```
[broccoli-persistent-filter:Babel > [Babel: ember-this-fallback]: Babel: ember-this-fallback] was configured to `throwUnlessParallelizable` and was unable to parallelize a plugin.
```

This commit updates the plugin definition to be compatible with parallel transformations, and adds the `throwUnlessParallelizable` flag to the dummy app. This change also seems to resolve the 'opting out of caching' error which previously required a `baseDir` hack.

More info on parallelization at https://github.com/babel/ember-cli-babel#parallel-builds and https://github.com/babel/broccoli-babel-transpiler#parallel-transpilation